debug physical test #4

+++ ports/nrf52840/core/hardware/ads1015/ads1015.cpp
@@ -35,7 +35,7 @@ void ads1015LL::send_command_conf(uint8_t *command)
 }
 
 
-void ads1015LL::read(double& digital_value) {
+void ads1015LL::read(int& digital_value) {
     //start trigger
     //code gpio trig ;
     //int flag = 1;
